+++ api_v3/lib/types/responseProfile/KalturaResponseProfileCacheRecalculateOptions.php
@@ -6,30 +6,30 @@
 class KalturaResponseProfileCacheRecalculateOptions extends KalturaObject
 {
 	/**
-	 * Unique identifier of object type and id
-	 * 
-	 * @var string
-	 */
-	public $objectKey;
-	
-	/**
-	 * First id to recalculate
-	 * 
-	 * @var string
-	 */
-	public $startKeyId;
-	
-	/**
-	 * Last id to recalculate
-	 * 
-	 * @var string
-	 */
-	public $endKeyId;
-	
-	/**
 	 * Maximum number of keys to recalculate
 	 * 
 	 * @var int
 	 */
 	public $limit;
+	
+	/**
+	 * Class name
+	 * @var string
+	 */
+	public $cachedObjectType;
+	
+	/**
+	 * @var string
+	 */
+	public $objectId;
+	
+	/**
+	 * @var string
+	 */
+	public $startDocId;
+	
+	/**
+	 * @var string
+	 */
+	public $endDocId;
 }